+++ docs/changelog.md
@@ -11,7 +11,7 @@ category: Administration
 
 ### [Unreleased]
 
-* ...
+* feat: ensure compatibility with Jira 9.7.0, 9.4.4 and 8.20.20
 
 ### [23.03.4] - 2023-03-29